Ginormous refactor to remove all redundant code
