+++ src/superphot_plus/priors/__init__.py
@@ -1,0 +1,3 @@
+from .generate_prior import generate_priors
+
+__all__ = ["generate_priors",]
