+++ src/superphot_plus/priors/priors_ZTF_r.csv
@@ -1,0 +1,8 @@
+param,min,max,mean,stddev,logged,relative_op
+A,-0.3,0.5,0.0957,0.0575,True,
+beta,-0.01,0.03,0.00833,0.00385,False,
+gamma,0.0,3.5,1.4258,0.3079,True,
+t_0,-100.0,30.0,-17.878,9.916,False,
+tau_rise,-2.0,4.0,0.6664,0.4250,True,
+tau_fall,0.0,4.0,1.5261,0.3037,True,
+extra_sigma,-3.0,-0.8,-1.6629,0.3378,True,
