+++ src/superphot_plus/samplers/superphot_sampler.py
@@ -1,69 +1,54 @@
 from typing import Optional
 
 import numpy as np
-from snapi import Sampler
 
-from superphot_plus.surveys.fitting_priors import MultibandPriors
-from superphot_plus.surveys.surveys import Survey
+from snapi.analysis import Sampler, SamplerPrior
 from superphot_plus.utils import flux_model
 
 class SuperphotSampler(Sampler):
     """Subclass of SNAPI Sampler for Superphot+."""
     def __init__(
         self,
+        priors: SamplerPrior,
         *args,
-        priors: MultibandPriors=Survey.ZTF().priors,
         **kwargs
     ):
         super().__init__()
         self._nparams = 6
-        self._all_priors = priors.to_numpy().T
-        self._ref_band = priors.reference_band
-        self._unique_bands = priors.ordered_bands
-        self._ref_band_idx = np.argmax(self._unique_bands == self._ref_band)
-        self._start_idx = 7 * self._ref_band_idx
-        self._is_fitted = False
+        self._priors = priors
+        self._params = self._priors.dataframe['param'].to_numpy()
+        self._unique_bands = []
+        for c in self._params:
+            if c[0] == 'A':
+                self._unique_bands.append(c[2:])
+        self._base_params = []
+        for c in self._params:
+            if self._unique_bands[0] in c:
+                self._base_params.append(c.replace("_" +self._unique_bands[0], ""))
         self.result = None
 
-
-    def _create_param_names(self):
-        """Creates the parameter names."""
-        param_names = ['log_A', 'beta', 'log_gamma', 't0', 'log_tau_rise', 'log_tau_fall', 'log_extra_sigma']
-        for band in self._unique_bands:
-            if band == self._ref_band:
-                continue
-            param_names += [
-                f'log_A_{band}', f'beta_{band}', f'log_gamma_{band}', f't0_{band}',
-                f'log_tau_rise_{band}', f'log_tau_fall_{band}', f'log_extra_sigma_{band}'
-            ]
-        return param_names
-
+    def _reformat_cube(self, cube):
+        """Reformat cube based on self._param_map"""
+        return cube[self._param_map]
+            
     def _eff_variance(self, X):
         """Calculates the effective variance of the model."""
-        log_extra_sigma_arr = np.repeat(self.result.fit_parameters['log_extra_sigma'].to_numpy()[:,np.newaxis], X.shape[0], axis=1)
-        
-        for ordered_band in self._unique_bands:
-            if ordered_band == self._ref_band:
-                continue
-            log_extra_sigma_arr_band = np.repeat(self.result.fit_parameters[f'log_extra_sigma_{ordered_band}'].to_numpy()[:,np.newaxis], X.shape[0], axis=1)
-            log_extra_sigma_arr[:,X[:,1] == ordered_band] += log_extra_sigma_arr_band[:,X[:,1] == ordered_band]
-        
-        return X[:,2:3].T.astype(np.float32)**2 + (10**log_extra_sigma_arr)**2
+        fit_param_numpy = self.result.fit_parameters[self._params].to_numpy().T # each entry is a parameter
+        extra_sigma_arr = self._reformat_cube(fit_param_numpy)[-1] # (num_times, num_fits)
+        return X[:,2:3].T.astype(np.float32)**2 + (extra_sigma_arr.T)**2
 
     def predict(self, X, num_fits=None):
         """Predicts the flux of a light curve using the model."""
+        fit_param_numpy = self.result.fit_parameters[self._params].to_numpy().T # each entry is a parameter
+        cube = self._reformat_cube(fit_param_numpy) # (num_params, num_times, num_fits)
         _, val_x = super().predict(X)
         if num_fits:
             return flux_model(
-                self.result.fit_parameters.to_numpy()[:num_fits],
-                val_x[:, 0].astype(np.float32), val_x[:, 1],
-                self._unique_bands,
-                self._ref_band
+                cube[:,:,:num_fits],
+                val_x[:, 0].astype(np.float32), val_x[:, 1]
             ), val_x
 
         return flux_model(
-            self.result.fit_parameters.to_numpy(),
-            val_x[:, 0].astype(np.float32), val_x[:, 1],
-            self._unique_bands,
-            self._ref_band
+            cube,
+            val_x[:, 0].astype(np.float32), val_x[:, 1]
         ), val_x
