--- src/superphot_plus/surveys/fitting_priors.py
@@ -1,137 +1,0 @@
-"""Data class for per-band curve priors"""
-
-from dataclasses import dataclass, field
-from typing import Dict
-
-import numpy as np
-
-
-@dataclass
-class PriorFields:
-    """Holder for per-parameter field characterization"""
-
-    clip_a: float = 0
-    clip_b: float = 0
-    mean: float = 0
-    std: float = 0
-
-    def to_numpy(self):
-        """Fields as a length-4 numpy array."""
-        return np.array([self.clip_a, self.clip_b, self.mean, self.std])
-
-
-@dataclass
-class CurvePriors:
-    """Set of priors for fitting a single curve."""
-
-    amp: PriorFields = None
-    beta: PriorFields = None
-    gamma: PriorFields = None
-    t_0: PriorFields = None
-    tau_rise: PriorFields = None
-    tau_fall: PriorFields = None
-    extra_sigma: PriorFields = None
-
-    def __post_init__(self):
-        """Additional logic to coerce string dictionaries into the appropriate
-        data type."""
-        if isinstance(self.amp, dict):
-            self.amp = PriorFields(**self.amp)
-            self.beta = PriorFields(**self.beta)
-            self.gamma = PriorFields(**self.gamma)
-            self.t_0 = PriorFields(**self.t_0)
-            self.tau_rise = PriorFields(**self.tau_rise)
-            self.tau_fall = PriorFields(**self.tau_fall)
-            self.extra_sigma = PriorFields(**self.extra_sigma)
-
-    def to_numpy(self):
-        """Fields as a 7x4 numpy array"""
-        return [
-            self.amp.to_numpy(),
-            self.beta.to_numpy(),
-            self.gamma.to_numpy(),
-            self.t_0.to_numpy(),
-            self.tau_rise.to_numpy(),
-            self.tau_fall.to_numpy(),
-            self.extra_sigma.to_numpy(),
-        ]
-
-
-@dataclass
-class MultibandPriors:
-    """Set of per-band curve priors"""
-
-    bands: Dict[str, CurvePriors] = field(default_factory=dict)
-    """Per-band curve priors."""
-    band_order: str = "ugrizy"
-    """Ordering of bands."""
-    reference_band: str = "r"
-    """Reference band."""
-    name: str = ""
-    """Name of survey."""
-
-    def __post_init__(self):
-        """Additional logic to coerce string dictionaries into the appropriate
-        data type."""
-        tmp_directory = self.bands.copy()
-        for band, curve_priors in self.bands.items():
-            if isinstance(curve_priors, dict):
-                tmp_directory[f"{self.name}_{band}"] = CurvePriors(**curve_priors)
-        self.bands = tmp_directory
-        self.reference_band = f"{self.name}_{self.reference_band}"
-
-    @property
-    def ordered_bands(self):
-        """Returns included bands in band_order."""
-        bands_ordered = []
-        for band in self.band_order:
-            if f"{self.name}_{band}" in self.bands:
-                bands_ordered.append(f"{self.name}_{band}")
-
-        return np.array(bands_ordered)
-
-    @property
-    def ref_band_index(self):
-        """Returns index of reference band."""
-        return np.argmax(self.ordered_bands == self.reference_band)
-    
-    @property
-    def aux_bands(self):
-        """Returns auxilliary (non-reference) bands in band_order."""
-        bands_ordered = []
-        for band in self.band_order:
-            if f"{self.name}_{band}" in self.bands and f"{self.name}_{band}" != self.reference_band:
-                bands_ordered.append(band)
-
-        return np.array(bands_ordered)
-
-    def filter_by_band(self, band_list, in_place=True):
-        """Return MultibandPriors object with only some bands.
-        """
-        assert self.reference_band[-1] in band_list
-        
-        bands_filtered = {band: self.bands[band] for band in band_list}
-        band_order = ""
-        for band in self.band_order:
-            if band in band_list:
-                band_order += band
-        
-        if in_place:
-            self.bands = bands_filtered
-            self.band_order = band_order
-            return self
-        
-        return MultibandPriors(
-            bands={b[-1]: bands_filtered[b] for b in bands_filtered},
-            band_order=band_order,
-            reference_band=self.reference_band[-1]
-        )        
-        
-    def to_numpy(self):
-        """Fields as a (7*bands)x4 numpy array"""
-        priors = []
-        for band in self.band_order:
-            if f"{self.name}_{band}" in self.bands:
-                priors.append(self.bands[f"{self.name}_{band}"].to_numpy())
-
-        return np.concatenate(priors)
