--- src/superphot_plus/surveys/ztf.yaml
@@ -1,80 +1,0 @@
-name: ZTF
-wavelengths:
-  r: 6173.23
-  g: 4741.64
-priors:
-  band_order: rg
-  reference_band: r
-  bands:
-    r:
-      amp:
-        clip_a: -0.3
-        clip_b: 0.5
-        mean: 0.0957
-        std: 0.0575
-      beta:
-        clip_a: -0.01
-        clip_b: 0.03
-        mean: 0.00833
-        std: 0.00385
-      gamma:
-        clip_a: 0.0
-        clip_b: 3.5
-        mean: 1.4258
-        std: 0.3079
-      t_0:
-        clip_a: -100.0
-        clip_b: 30.0
-        mean: -17.878
-        std: 9.916
-      tau_rise:
-        clip_a: -2.0
-        clip_b: 4.0
-        mean: 0.6664
-        std: 0.4250
-      tau_fall:
-        clip_a: 0.0
-        clip_b: 4.0
-        mean: 1.5261
-        std: 0.3037
-      extra_sigma:
-        clip_a: -3.0
-        clip_b: -0.8
-        mean: -1.6629
-        std: 0.3378
-    g:
-      amp:
-        clip_a: -1.0
-        clip_b: 1.0
-        mean: -0.0766
-        std: 0.1096
-      beta:
-        clip_a: -0.01
-        clip_b: 0.03
-        mean: 0.0
-        std: 0.004
-      gamma:
-        clip_a: -1.5
-        clip_b: 1.5
-        mean: -0.0452
-        std: 0.1705
-      t_0:
-        clip_a: -50
-        clip_b: 30
-        mean: -3.4065
-        std: 4.366
-      tau_rise:
-        clip_a: -1.5
-        clip_b: 1.5
-        mean: -0.1510
-        std: 0.1927
-      tau_fall:
-        clip_a: -1.5
-        clip_b: 1.5
-        mean: -0.1486
-        std: 0.2642
-      extra_sigma:
-        clip_a: -1.5
-        clip_b: 1.0
-        mean: -0.1509
-        std:  0.2542
